Move assets to their own file
