+++ src/Config/scenes.js
@@ -6,6 +6,7 @@ import TitleScene from './../Scenes/TitleScene';
 import OptionsScene from './../Scenes/OptionsScene';
 import CreditsScene from './../Scenes/CreditsScene';
 import UIScene from './../Scenes/UIScene';
+import GameOverScene from './../Scenes/GameOverScene';
 
 const scenes = [
   { key: 'Boot', sceneConfig: BootScene},
@@ -15,7 +16,8 @@ const scenes = [
   { key: 'Credits', sceneConfig: CreditsScene},
   { key: 'Game', sceneConfig: GameScene},
   { key: 'Battle', sceneConfig: BattleScene},
-  { key: 'UIScene', sceneConfig: UIScene}
+  { key: 'UIScene', sceneConfig: UIScene},
+  { key: 'GameOver', sceneConfig: GameOverScene}
 ];
 
 export default scenes;
