+++ src/Scenes/GameScene.js
@@ -1,5 +1,6 @@
 import 'phaser';
 import { loadAssets, createAnims } from './../Objects/Utilities';
+import assets from './../Config/assets';
 
 export default class GameScene extends Phaser.Scene {
   constructor() {
@@ -7,46 +8,6 @@ export default class GameScene extends Phaser.Scene {
   }
 
   preload() {
-    const assets = [{
-        key: 'tiles',
-        path: 'assets/map/spritesheet.png',
-        type: 'image'
-      },
-      {
-        key: 'map',
-        path: 'assets/map/map.json',
-        type: 'tilemapTiledJSON'
-      },
-      {
-        key: 'ork',
-        path: 'assets/ork.png',
-        type: 'image'
-      },
-      {
-        key: 'dragonorrange',
-        path: 'assets/dragonorrange.png',
-        type: 'image'
-      },
-      {
-        key: 'player',
-        path: 'assets/RPG_assets.png',
-        opts: {
-          frameWidth: 32,
-          frameHeight: 32
-        },
-        type: 'spritesheet'
-      },
-      {
-        key: 'life_bar',
-        path: 'assets/life.png',
-        opts: {
-          frameWidth: 214,
-          frameHeight: 28
-        },
-        type: 'spritesheet'
-      }
-    ];
-
     loadAssets(this, assets);
     this.sys.events.on('update health points', this.updateHealthPoints, this);
   }
